+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/models/chores/Chore.kt
@@ -2,9 +2,11 @@ package com.altmann.choresmanager.models.chores
 
 import androidx.compose.ui.graphics.Color
 import com.altmann.choresmanager.models.chores.Priority
+import kotlinx.datetime.DateTimePeriod
 import kotlinx.datetime.DayOfWeek
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.LocalTime
+import kotlin.math.abs
 
 open class Chore(
     val choreId: String,
@@ -22,7 +24,7 @@ open class Chore(
     val color: Color = Color.DarkGray
 ) {
     open fun duration(): Int {
-        return endTime.hour - startTime.hour
+        return (endTime.hour * 60 + endTime.minute) - (startTime.hour * 60 + startTime.minute)
     }
 
     open fun finishChore() {